+++ test/mixed.ts
@@ -673,7 +673,7 @@ describe('Mixed Types ', () => {
   it('concat should carry over transforms', async () => {
     let inst = string().trim();
 
-    await expect(inst.concat(string().min(4)).cast(' hello  ')).toBe('hello');
+    expect(inst.concat(string().min(4)).cast(' hello  ')).toBe('hello');
 
     await expect(inst.concat(string().min(4)).isValid(' he  ')).resolves.toBe(
       false,
@@ -957,7 +957,7 @@ describe('Mixed Types ', () => {
         foo: array(number().integer()).required(),
         bar: string()
           .max(2)
-          .default(()=> 'a')
+          .default(() => 'a')
           .meta({ input: 'foo' })
           .label('str!')
           .oneOf(['a', 'b'])
@@ -966,7 +966,7 @@ describe('Mixed Types ', () => {
             is: 'entered',
             then: (s) => s.defined(),
           }),
-        baz: tuple([string(), number()])
+        baz: tuple([string(), number()]),
       });
     });
 
@@ -1070,7 +1070,7 @@ describe('Mixed Types ', () => {
                 oneOf: [],
                 notOneOf: [],
                 tests: [],
-              }
+              },
             ],
           },
         },
@@ -1182,7 +1182,7 @@ describe('Mixed Types ', () => {
                 oneOf: [],
                 notOneOf: [],
                 tests: [],
-              }
+              },
             ],
           },
         },
@@ -1215,6 +1215,40 @@ describe('Mixed Types ', () => {
       });
 
       await expect(inst.isValid({ prop: 'prop value' })).resolves.toBe(true);
+    });
+  });
+
+  describe('description options', () => {
+    const schema = object({
+      name: string(),
+      type: bool(),
+      fancy: string()
+        .label('bad label')
+        .when('type', {
+          is: true,
+          then: (schema) => schema.required().label('good label'),
+          otherwise: (schema) => schema.label('default label'),
+        }),
+    });
+
+    it('should pass options', async () => {
+      expect(
+        // @ts-ignore
+        schema.fields.fancy.describe({ parent: { type: true } }).label,
+      ).toBe('good label');
+      expect(
+        // @ts-ignore
+        schema.fields.fancy.describe({ parent: { type: true } }).optional,
+      ).toBe(false);
+
+      expect(
+        // @ts-ignore
+        schema.fields.fancy.describe({ parent: { type: false } }).label,
+      ).toEqual('default label');
+      expect(
+        // @ts-ignore
+        schema.fields.fancy.describe({ parent: { type: false } }).optional,
+      ).toBe(true);
     });
   });
 });